refactor SceneGIContext
